Add "arguments" keyword processing into the MATLAB lexer
"arguments" is context dependent in MATLAB. If it is used right
after a function declaration (commented lines don't count), it
denotes the arguments code block describing the restrictions on
the function's arguments. If it is used elsewhere, it's just a
regular identifier.
The code in this commit marks "arguments" as a keyword opening
the folded code block only if it is really a keyword, and avoids
folding in cases where "arguments" is used as an identifier.